Bug: 236760014
Merged-In: I02db31e630f62a941dca13322cdead76442cbf99

+++ cpp/watchdog/server/tests/UidProcStatsCollectorTest.cpp
@@ -481,7 +481,7 @@ TEST(UidProcStatsCollectorTest, TestUidProcStatsCollectorContentsFromDevice) {
     const auto& processStats = collector.deltaStats();
 
     // The below check should pass because there should be at least one process.
-    EXPECT_GT(processStats.size(), 0);
+    EXPECT_GT(processStats.size(), static_cast<size_t>(0));
 }
 
 }  // namespace watchdog
